reverting to last working img-opt.js

+++ utils/img-opt.js
@@ -25,13 +25,15 @@ async function writeThumbs(dir, sizes, name, format, width, height) {
     const sizeConf = sizes[size];
     const outputFn = join(dir, 'thumbs', `${name}-${size}.webp`);
     const imagePath = join(dir, name + '.' + format);
-    
+    if (!(await exists(imagePath))) {
+      console.log(`${imagePath} already exists, skipping ...`);
+      continue;
+    }
     const image = await Deno.readFile(imagePath);
     const resized = await resize(image, sizeConf.width, Math.round(height / (width / sizeConf.width)));
-    const webp = await mod.ImageToWebP(resized, { quality: 80 });
-    
+    const webp = await mod.imageToWebP(resized);
     await Deno.writeFile(outputFn, webp);
-    console.log(`Thumbnail generated: ${outputFn}`); // Log the generated thumbnail
+    console.log(`thumbnail written: ${outputFn}`);
   }
 }
 
@@ -40,28 +42,24 @@ async function optimizeDir(dir, sizes) {
   await checkThumbs(dir);
 
   for await (const f of Deno.readDir(dir)) {
-    const [name] = f.name.split('.');
-    const formats = ['jpg', 'jpeg', 'png', 'JPG', 'JPEG', 'PNG'];
-    let foundImage = false;
-
-    for (const ext of formats) {
-      const imagePath = join(dir, `${name}.${ext}`);
-      if (await exists(imagePath)) {
-        console.log(`Processing: ${name}.${ext}`);
-        foundImage = true;
-
-        const explain = await run(`identify ${imagePath}`);
-        const [_, format, resolution] = explain.split(' ');
-        const [width, height] = resolution.split('x').map(c => Number(c));
-
-        await writeThumbs(dir, sizes, name, ext, width, height);
-        break; // Exit the loop once the first found image is processed
-      }
+    const [name, ext] = f.name.split('.');
+    if (!['jpg', 'jpeg', 'png'].includes(ext)) {
+      continue;
     }
 
-    if (!foundImage) {
-      console.log(`No valid image found for: ${name}, please upload it to /src/people/_images/ folder ...`);
+    const imagePath = join(dir, f.name);
+    if (!(await exists(imagePath))) {
+      console.log(`${imagePath} already exists, skipping ...`);
+      continue;
     }
+
+    console.log(`processing: ${name}`);
+
+    const explain = await run(`identify ${join(dir, f.name)}`);
+    const [_, format, resolution] = explain.split(' ');
+    const [width, height] = resolution.split('x').map(c => Number(c));
+
+    await writeThumbs(dir, sizes, name, format, width, height);
   }
 }
